# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Voxara API — music-driven language & pronunciation tutor
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Track {
  trackId: number;
  commontrackId: number;
  trackName: string;
  artistName: string;
  /** @nullable */
  albumName?: string | null;
  /** @nullable */
  albumCoverUrl?: string | null;
  hasLyrics: boolean;
  /** Whether word-level timing is available (best experience). */
  hasRichsync: boolean;
  /** Whether line-level subtitle timing is available. Used to predict the "line practice" tier before a session is resolved.
   */
  hasSubtitles: boolean;
  hasTranslation: boolean;
}

export interface BreakoutTrack {
  trackId: number;
  commontrackId: number;
  trackName: string;
  artistName: string;
  /** @nullable */
  albumName?: string | null;
  /** @nullable */
  albumCoverUrl?: string | null;
  hasLyrics: boolean;
  hasRichsync: boolean;
  hasSubtitles: boolean;
  hasTranslation: boolean;
  /** Songstats-derived breakout/velocity score (higher = more momentum). */
  velocityScore: number;
}

export interface TrackStats {
  /** Whether a matching track was found on Songstats. */
  found: boolean;
  /**
     * Total Spotify streams.
     * @nullable
     */
  spotifyStreams?: number | null;
  /**
     * Spotify popularity index (0-100).
     * @nullable
     */
  spotifyPopularity?: number | null;
  /**
     * Current Spotify playlist placements.
     * @nullable
     */
  playlists?: number | null;
  /**
     * Total Shazams.
     * @nullable
     */
  shazams?: number | null;
  /**
     * Total TikTok video views using the track.
     * @nullable
     */
  tiktokViews?: number | null;
  /**
     * Link to the track's Songstats profile.
     * @nullable
     */
  songstatsUrl?: string | null;
}

export interface LyricWord {
  text: string;
  /** Word start time in seconds. */
  start: number;
  /** Word end time in seconds. */
  end: number;
}

export interface LyricLine {
  index: number;
  /** Line start time in seconds. */
  ts: number;
  /** Line end time in seconds. */
  te: number;
  /** The full original line text. */
  text: string;
  words: LyricWord[];
  /**
     * Translated text for this line, if available.
     * @nullable
     */
  translation?: string | null;
}

/**
 * Timing granularity available for this track — "word" (per-word richsync), "line" (line-level subtitles), or "none" (plain lyrics only, Practice unavailable).

 */
export type TrackSessionSyncLevel = typeof TrackSessionSyncLevel[keyof typeof TrackSessionSyncLevel];


export const TrackSessionSyncLevel = {
  word: 'word',
  line: 'line',
  none: 'none',
} as const;

export interface TrackSession {
  track: Track;
  lines: LyricLine[];
  /**
     * Plain-text lyrics fallback when richsync is unavailable.
     * @nullable
     */
  plainLyrics?: string | null;
  /** Mandatory Musixmatch copyright notice to render alongside lyrics. */
  copyright: string;
  /** Whether word-level timing is available (syncLevel == "word"). */
  hasRichsync: boolean;
  hasTranslation: boolean;
  /** Timing granularity available for this track — "word" (per-word richsync), "line" (line-level subtitles), or "none" (plain lyrics only, Practice unavailable).
   */
  syncLevel: TrackSessionSyncLevel;
  targetLanguage: string;
}

export type GradedWordStatus = typeof GradedWordStatus[keyof typeof GradedWordStatus];


export const GradedWordStatus = {
  match: 'match',
  substituted: 'substituted',
  missing: 'missing',
} as const;

export interface GradedWord {
  expected: string;
  status: GradedWordStatus;
  /**
     * What the learner actually said for this position, if any.
     * @nullable
     */
  actual?: string | null;
}

export interface GradingResult {
  /** Matched words / total expected words, as a percentage. */
  scorePercent: number;
  /** Full transcription of the learner's attempt. */
  transcript: string;
  words: GradedWord[];
  /** Words the learner added that were not expected. */
  extraWords: string[];
}

export type GetFeaturedTracksParams = {
/**
 * Target language code used to indicate translation availability.
 */
selected_language?: string;
};

export type SearchTracksParams = {
/**
 * Track title.
 */
q_track?: string;
/**
 * Artist name.
 */
q_artist?: string;
/**
 * Free-text query (artist and title combined).
 */
q?: string;
};

export type IdentifyTrackParams = {
/**
 * A snippet of lyrics the learner remembers.
 */
lyric: string;
};

export type DiscoverByMoodParams = {
/**
 * The mood to discover tracks for.
 */
mood: DiscoverByMoodMood;
};

export type DiscoverByMoodMood = typeof DiscoverByMoodMood[keyof typeof DiscoverByMoodMood];


export const DiscoverByMoodMood = {
  heartbreak: 'heartbreak',
  hype: 'hype',
  nostalgic: 'nostalgic',
  romantic: 'romantic',
  hopeful: 'hopeful',
  chill: 'chill',
} as const;

export type GetTrendingTracksParams = {
/**
 * Target language code used to indicate translation availability.
 */
selected_language?: string;
};

export type GetTrackStatsParams = {
/**
 * Track title.
 */
trackName: string;
/**
 * Artist name.
 */
artistName: string;
};

export type GetTrackSessionParams = {
/**
 * Musixmatch track_id.
 */
trackId: number;
/**
 * Target language code for translation (e.g. "en").
 */
selected_language?: string;
};