# lib/api-zod/src/generated/types/breakoutTrack.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Voxara API — music-driven language & pronunciation tutor
 * OpenAPI spec version: 0.1.0
 */

export interface BreakoutTrack {
  trackId: number;
  commontrackId: number;
  trackName: string;
  artistName: string;
  /** @nullable */
  albumName?: string | null;
  /** @nullable */
  albumCoverUrl?: string | null;
  hasLyrics: boolean;
  hasRichsync: boolean;
  hasSubtitles: boolean;
  hasTranslation: boolean;
  /** Songstats-derived breakout/velocity score (higher = more momentum). */
  velocityScore: number;
}